requirements.txt
numpy
pandas
scikit-learn
pillow
matplotlib